worker/requirements.txt
grpcio
grpcio-tools
mlx-lm
